put the routes for the docs into its own module
